+++ Roguelike TBS/Assets/Scripts/Enemies/Enemy.cs
@@ -5,6 +5,7 @@ public class Enemy : MonoBehaviour {
     [SerializeField] protected AudioClip sfx = null;
     [SerializeField] protected float movementSpeed = 0f;
     [SerializeField] int moneyDropped = 1;
+    [SerializeField] protected HealthBar healthBar;
 
     /// CACHE
     protected GameObject player;
@@ -25,7 +26,7 @@ public class Enemy : MonoBehaviour {
 
     private void OnTriggerEnter2D(Collider2D other) {
         if (other.gameObject.tag == "Weapon") {
-            gameObject.GetComponent<Animator>().SetTrigger("Hit");
+            animator.SetTrigger("Hit");
             GameObject wielder = other.gameObject.GetComponent<Weapon>().GetWielder().gameObject;
             var damageTaken = wielder.GetComponent<BaseStats>().GetStat(Stat.Attack) - gameObject.GetComponent<BaseStats>().GetStat(Stat.Defense);
             Debug.Log("Damage being dealt is " + damageTaken);
@@ -39,7 +40,7 @@ public class Enemy : MonoBehaviour {
 
     protected void OnCollisionEnter2D(Collision2D other) {
         if (other.gameObject.tag == "Item") {
-            gameObject.GetComponent<Animator>().SetTrigger("Hit");
+            animator.SetTrigger("Hit");
             var damageTaken = other.gameObject.GetComponent<ThrownItem>().GetDamage() - gameObject.GetComponent<BaseStats>().GetStat(Stat.Defense);
 
             health.TakeDamage(
@@ -50,7 +51,7 @@ public class Enemy : MonoBehaviour {
                 AudioSource.PlayClipAtPoint(sfx, transform.position);
             }
         } else if (other.gameObject.tag == "Magic") {
-            gameObject.GetComponent<Animator>().SetTrigger("Hit");
+            animator.SetTrigger("Hit");
             var damageTaken = other.gameObject.GetComponent<MagicMissile>().GetDamage() - gameObject.GetComponent<BaseStats>().GetStat(Stat.Defense);
 
             health.TakeDamage(